Bug fix
Bugged person on first level
Check for Dummy on death touch

+++ Assets/Scripts/Core/Person/Other/PlayablePersonDeathTouch.cs
@@ -12,7 +12,7 @@ namespace Core.Person.Other
         private void OnTriggerEnter(Collider other)
         {
             if (!other.TryGetComponent<PlayablePerson>(out var person)) return;
-            
+            if (person.IsDummy) return;
             person.Kill();
         }
     }
